+++ Taarafo.Portal.Web.Tests.Unit/Services/Foundations/Posts/PostServiceTests.cs
@@ -10,7 +10,7 @@ using System.Linq.Expressions;
 using System.Net.Http;
 using Moq;
 using RESTFulSense.Exceptions;
-using Taarafo.Portal.Web.Brokers.API;
+using Taarafo.Portal.Web.Brokers.Apis;
 using Taarafo.Portal.Web.Brokers.Loggings;
 using Taarafo.Portal.Web.Models.Posts;
 using Taarafo.Portal.Web.Services.Foundations.Posts;
