+++ Taarafo.Portal.Web/Services/Foundations/Posts/PostService.cs
@@ -6,7 +6,7 @@
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
-using Taarafo.Portal.Web.Brokers.API;
+using Taarafo.Portal.Web.Brokers.Apis;
 using Taarafo.Portal.Web.Brokers.Loggings;
 using Taarafo.Portal.Web.Models.Posts;
